Support full RMSNorm corpus with 64-bit stores and chunked host checks

+++ tests/softmax_rmsnorm_950/inputs/rmsnorm_affine_store.cuh
@@ -17,6 +17,7 @@ limitations under the License.
 // Conversion input mirroring the forward adapter in
 // oneflow/user/kernels/rms_norm_gpu_kernel.cu.  It intentionally contains no
 // Ascify target markers; the dav-c310 recipe must prove and add them.
+// Linear offsets use 64 bits, matching the standalone A800 benchmark adapter.
 #ifndef ASCIFY950_RMSNORM_AFFINE_STORE_INPUT_CUH_
 #define ASCIFY950_RMSNORM_AFFINE_STORE_INPUT_CUH_
 
@@ -28,15 +29,15 @@ namespace ascify950_recipe_fixture {
 
 template<typename SRC, typename DST, bool affine>
 struct RmsNormAffineStore {
-  RmsNormAffineStore(DST* dst, const DST* weight, int32_t row_size)
+  RmsNormAffineStore(DST* dst, const DST* weight, int64_t row_size)
       : dst(dst), weight(weight), row_size(row_size) {}
 
   template<int N>
-  __device__ void store(const SRC* src, int32_t row, int32_t col) {
+  __device__ void store(const SRC* src, int64_t row, int64_t col) {
     oneflow::cuda::layer_norm::Pack<DST, N> dst_pack;
     oneflow::cuda::layer_norm::Pack<DST, N> weight_pack;
-    const int32_t offset = (row * row_size + col) / N;
-    const int32_t weight_offset = col / N;
+    const int64_t offset = (row * row_size + col) / N;
+    const int64_t weight_offset = col / N;
     if (affine) {
       weight_pack.storage =
           *(reinterpret_cast<const oneflow::cuda::layer_norm::PackType<DST, N>*>(
@@ -58,7 +59,7 @@ struct RmsNormAffineStore {
 
   DST* dst;
   const DST* weight;
-  int32_t row_size;
+  int64_t row_size;
 };
 
 }  // namespace ascify950_recipe_fixture
